Maintains aspect ratio on playback. Opportunistically moves ProgressBar to its own file
