+++ packages/teleport/src/Player/DesktopPlayer.tsx
@@ -1,17 +1,14 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 
 import styled from 'styled-components';
-import { throttle } from 'lodash';
-import { dateToUtc } from 'shared/services/loc';
-import { format } from 'date-fns';
 
 import cfg from 'teleport/config';
-import { PlayerClient, PlayerClientEvent } from 'teleport/lib/tdp';
+import { PlayerClient } from 'teleport/lib/tdp';
 import { PngFrame, ClientScreenSpec } from 'teleport/lib/tdp/codec';
 import { getAccessToken, getHostName } from 'teleport/services/api';
 import TdpClientCanvas from 'teleport/components/TdpClientCanvas';
 
-import ProgressBar from './ProgressBar';
+import { ProgressBarDesktop } from './ProgressBar';
 
 export const DesktopPlayer = ({
   sid,
@@ -30,6 +27,8 @@ export const DesktopPlayer = ({
 
   return (
     <StyledPlayer>
+      {/* NOTE: tdpCliOnClientScreenSpec depends on the order of elements within StyledPlayer.
+      Be sure to update it if the elements within are changed. */}
       <TdpClientCanvas
         tdpCli={playerClient}
         tdpCliOnPngFrame={tdpCliOnPngFrame}
@@ -37,81 +36,13 @@ export const DesktopPlayer = ({
         onContextMenu={() => true}
         // overflow: 'hidden' is needed to prevent the canvas from outgrowing the container due to some weird css flex idiosyncracy.
         // See https://gaurav5430.medium.com/css-flex-positioning-gotchas-child-expands-to-more-than-the-width-allowed-by-the-parent-799c37428dd6.
-        style={{ display: 'flex', flexGrow: 1, overflow: 'hidden' }}
+        style={{
+          alignSelf: 'center',
+          overflow: 'hidden',
+        }}
       />
       <ProgressBarDesktop playerClient={playerClient} durationMs={durationMs} />
     </StyledPlayer>
-  );
-};
-
-export const ProgressBarDesktop = (props: {
-  playerClient: PlayerClient;
-  durationMs: number;
-}) => {
-  const { playerClient, durationMs } = props;
-
-  const toHuman = (currentMs: number) => {
-    return format(dateToUtc(new Date(currentMs)), 'mm:ss');
-  };
-
-  const [state, setState] = useState({
-    max: durationMs,
-    min: 0,
-    current: 0, // the recording always starts at 0 ms
-    time: toHuman(0),
-    isPlaying: true, // determines whether play or pause symbol is shown
-  });
-
-  useEffect(() => {
-    playerClient.addListener(PlayerClientEvent.TOGGLE_PLAY_PAUSE, () => {
-      // setState({...state, isPlaying: !state.isPlaying}) doesn't work because
-      // the listener is added when state == initialState, and that initialState
-      // value is effectively hardcoded into its logic.
-      setState(prevState => {
-        return { ...prevState, isPlaying: !prevState.isPlaying };
-      });
-    });
-
-    const throttledUpdateCurrentTime = throttle(
-      currentTimeMs => {
-        setState(prevState => {
-          return {
-            ...prevState,
-            current: currentTimeMs,
-            time: toHuman(currentTimeMs),
-          };
-        });
-      },
-      // Magic number to throttle progress bar updates so that the playback is smoother.
-      50
-    );
-
-    playerClient.addListener(
-      PlayerClientEvent.UPDATE_CURRENT_TIME,
-      currentTimeMs => throttledUpdateCurrentTime(currentTimeMs)
-    );
-
-    playerClient.addListener(PlayerClientEvent.SESSION_END, () => {
-      throttledUpdateCurrentTime.cancel();
-      // TODO(isaiah): Make this smoother
-      // https://github.com/gravitational/webapps/issues/579
-      setState(prevState => {
-        return { ...prevState, current: durationMs };
-      });
-    });
-
-    return () => {
-      throttledUpdateCurrentTime.cancel();
-      playerClient.nuke();
-    };
-  }, [playerClient]);
-
-  return (
-    <ProgressBar
-      {...state}
-      toggle={() => playerClient.togglePlayPause()}
-      move={() => {}}
-    />
   );
 };
 
@@ -141,6 +72,21 @@ const useDesktopPlayer = ({
     canvas: HTMLCanvasElement,
     spec: ClientScreenSpec
   ) => {
+    const styledPlayer = canvas.parentElement;
+    const progressBar = styledPlayer.children.item(1);
+
+    const fullWidth = styledPlayer.clientWidth;
+    const fullHeight = styledPlayer.clientHeight - progressBar.clientHeight;
+    const originalAspectRatio = spec.width / spec.height;
+    const currentAspectRatio = fullWidth / fullHeight;
+
+    if (originalAspectRatio > currentAspectRatio) {
+      // Use the full width of the screen and scale the height.
+      canvas.style.height = `${(fullWidth * spec.height) / spec.width}px`;
+    } else if (originalAspectRatio < currentAspectRatio) {
+      canvas.style.width = `${(fullHeight * spec.width) / spec.height}px`;
+    }
+
     canvas.width = spec.width;
     canvas.height = spec.height;
   };
@@ -155,6 +101,7 @@ const useDesktopPlayer = ({
 const StyledPlayer = styled.div`
   display: flex;
   flex-direction: column;
+  justify-content: center;
   width: 100%;
   height: 100%;
 `;
