+++ packages/teleport/src/Player/ProgressBar/index.tsx
@@ -16,6 +16,7 @@ limitations under the License.
 
 import ProgressBar from './ProgressBar';
 import ProgressBarTty from './ProgressBarTty';
+import { ProgressBarDesktop } from './ProgressBarDesktop';
 
 export default ProgressBar;
-export { ProgressBarTty };
+export { ProgressBarTty, ProgressBarDesktop };
